Feature added (Mocks): Several mocks to mock-up controls/permissions/clients
* CP_ANY: Modifications to design notes.
* CP_CLIENT: Added notion of `permissions' list.
* CP_PERMISSION: Rename to more appropriate name
* CP_PERMISSIONED_EV_TEXT_FIELD: While this will NOT end up being how
this works, it does quickly demonstrate use of EV in this mix.
* CP_PERMISSIONED_ITEM: Added creation procedure `make_with_client' and
`set_new_client_level' which allows changing of permission level
on-the-fly.
* TEST_SETs: Various tests of setting permissions on a control with a
client.
* MOCK_*: Various mock classes.

NOTE: The key is to get a matching UUID on the control and on the
permission, assign the permission to the Client and then give the Client
to the Control.