Features added (Registration): Builds a registry through a `put_unique'
* CP_ITEM_REGISTRY: The registry is now built through a feature that
ensures each item going into the registry is unique. This will correct
for a potential bug seen in an earlier check-in where I erroneously
included a GP_PERMISSION in the registry, which is a NO-NO!
* CP_PERMISSIONED_EV_TEXT_FIELD: Factored back `set_item_permission' to
a more appropriate level.
* MOCK_CARTOON_NAME_FIELD: Factored out `set_item_permission' to
CP_PERMISSIONED_EV_TEXT_FIELD.
* MOCK_ITEM_REGISTRY: Implements `build_items' instead of the `items'
hash directly, which allows for the call to `put_unique', which ensure
unique items are being added to the registry.
* MOCK_PHONE_NUMBER_FIELD: This gives us something else to add to the
registry with a different UUID, so we can test the `put_unique' and
`build_items' features (above).

NOTE: Overall, I am delighted with this library so far. The next step
will be to apply it to a working system. Perhaps to the eifwingen
project?